==> archives/FInal_presentation_02032023/1_DiabetesPrediction_DataProcessing_RawData.ipynb <==
### This data processing is run on the original raw dataset, no duplicates have been removed. (Duplicate data 'weighs' records heavier in the model.)

#import dependencies
import warnings
warnings.filterwarnings('ignore')
print('done')

#import dependencies
import numpy as np
import pandas as pd
import os
from pathlib import Path
from collections import Counter
print('done')

# import modeling dependencies
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn import linear_model
from sklearn.linear_model import LogisticRegression
from sklearn.feature_selection import SelectPercentile
from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import chi2 , f_classif
print('done')

#set file path
file_path3 = Path(r'resources\diabetes.csv')

#read file into memory
diabetes_df = pd.read_csv(file_path3)

#validate load
diabetes_df

#Dataframe Column_listing
diabetes_df.columns

#Determine Null Values 
nan_count = diabetes_df.isna().sum(axis = 1)
nan_count

#Check Data types - all fields should be numeric 
diabetes_df.dtypes

#Count number of each value in Diabetes 012 column
#0-no diabetes, 1-prediabetes 2-diabetes
#this is what we want our model to train on - recommend combining 1&2 into one value to make it binary 'get checked'
diabetes_df['Diabetes_012'].value_counts()

#percentage of each value in Diabetes 012 column
diabetes_df.Diabetes_012.value_counts(normalize=True)

#binning of Diabetes Status 
status_bin_range = [-1, 0, 4]

status_bins = [1, 2]

diabetes_df['Diabetes_Status'] = pd.cut(diabetes_df['Diabetes_012'], status_bin_range, labels=status_bins)

diabetes_df.Diabetes_Status.describe()

#examine columns 
diabetes_df.columns

#rearrange columns
diabetes_df = diabetes_df[['Diabetes_Status', 'HighBP', 'HighChol', 'CholCheck', 'BMI', 'Smoker',
       'Stroke', 'HeartDiseaseorAttack', 'PhysActivity', 'Fruits', 'Veggies',
       'HvyAlcoholConsump', 'AnyHealthcare', 'NoDocbcCost', 'GenHlth',
       'MentHlth', 'PhysHlth', 'DiffWalk', 'Sex', 'Age', 'Education', 'Income',
       'Diabetes_012']]

#confirm order 
diabetes_df.head()

#view info
diabetes_df.info()

diabetes_df['Diabetes_Status'].value_counts()

#Count number of each value in HighBP Column
# 0 - noHBP, 1 -  HBP 
diabetes_df['HighBP'].value_counts()


diabetes_df.info()

#percentage of each value 
diabetes_df.HighBP.value_counts(normalize=True)

#Count number of each value in HighChol Column
#0 = no high cholesterol 1 = high cholesterol
diabetes_df['HighChol'].value_counts()


#percentage
diabetes_df.HighChol.value_counts(normalize=True)

#Count number of each value in CholCheck Column
#0 = no cholesterol check in 5 years 1 = yes cholesterol check in 5 years
diabetes_df['CholCheck'].value_counts()


#percentage
diabetes_df.CholCheck.value_counts(normalize=True)

#Count number of each value in BMI Column
#calculate BMI by dividing weight in pounds (lb) by height in inches (in) squared and multiplying by a conversion factor of 703
#Will need to have user enter height and weight of subject 

diabetes_df['BMI'].value_counts()


#Determine bucketing for BMI values 
diabetes_df.BMI.describe()

#percentage
diabetes_df.BMI.value_counts(normalize=True)

# BMI Binning 
# https://qpp.cms.gov/docs/QPP_quality_measure_specifications/CQM-Measures/2019_Measure_128_MIPSCQM.pdf
# BMI 1 is less than 18.5
# BMI 2  is 18.5 to 24.9,
# BMI 3  is 25.0 to 29.9
# BMI 4 is 30.0 or higher

# define ranges
bmi_bin_range = [0, 18.5, 24.9, 29.9, 100]

#define bins 
bmi_bins = [1, 2, 3, 4]

#sort data into bins 
diabetes_df['BMI_Range'] = pd.cut(diabetes_df['BMI'], bmi_bin_range, labels=bmi_bins)

#describe model after binning 
diabetes_df.BMI_Range.describe()

#Validate columns post binning
diabetes_df.columns

#re-sort columns 
diabetes_df = diabetes_df[['Diabetes_Status', 'HighBP', 'HighChol', 'CholCheck', 'BMI_Range', 'Smoker',
       'Stroke', 'HeartDiseaseorAttack', 'PhysActivity', 'Fruits', 'Veggies',
       'HvyAlcoholConsump', 'AnyHealthcare', 'NoDocbcCost', 'GenHlth',
       'MentHlth', 'PhysHlth', 'DiffWalk', 'Sex', 'Age', 'Education', 'Income',
       'Diabetes_012', 'BMI']]

#view data 
diabetes_df.head()

#Count number of each value in Smoker Column
#Have you smoked at least 100 cigarettes in your entire life? [Note: 5 packs = 100 cigarettes] 0 = no 1 = yes
diabetes_df['Smoker'].value_counts()


#percentage
diabetes_df.Smoker.value_counts(normalize=True)

#Count number of each value in Stroke Column
#(Ever told) you had a stroke. 0 = no 1 = yes
diabetes_df['Stroke'].value_counts()

#Percentage 
diabetes_df.Stroke.value_counts(normalize=True)

#Count number of each value in HeartDiseaseorAttack Column
#coronary heart disease (CHD) or myocardial infarction (MI) 0 = no 1 = yes
diabetes_df['HeartDiseaseorAttack'].value_counts()


#percentage
diabetes_df.HeartDiseaseorAttack.value_counts(normalize=True)

#Count number of each value in PhysActivity Column
#physical activity in past 30 days - not including job 0 = no 1 = yes
diabetes_df['PhysActivity'].value_counts()


#percentage
diabetes_df.PhysActivity.value_counts(normalize=True)

#Count number of each value in Fruits Column
#Consume Fruit 1 or more times per day 0 = no 1 = yes
diabetes_df['Fruits'].value_counts()


#percentage
diabetes_df.Fruits.value_counts(normalize=True)

#Count number of each value in Veggies Column
#Consume Vegetables 1 or more times per day 0 = no 1 = yes
diabetes_df['Veggies'].value_counts()


#percentage
diabetes_df.Veggies.value_counts(normalize=True)

#Count number of each value in HvyAlcoholConsump Column
#Heavy drinkers (adult men having more than 14 drinks per week and adult women having more than 7 drinks per week) 0 = no 1 = yes
diabetes_df['HvyAlcoholConsump'].value_counts()

#percentage
diabetes_df.HvyAlcoholConsump.value_counts(normalize=True)

#Count number of each value in AnyHealthcare Column
#Have any kind of health care coverage, including health insurance, prepaid plans such as HMO, etc. 0 = no 1 = yes
diabetes_df['AnyHealthcare'].value_counts()

#percentage
diabetes_df.AnyHealthcare.value_counts(normalize=True)

#Count number of each value in NoDocbcCost Column
#Was there a time in the past 12 months when you needed to see a doctor but could not because of cost? 0 = no 1 = yes
diabetes_df['NoDocbcCost'].value_counts()

#percentage
diabetes_df.NoDocbcCost.value_counts(normalize=True)

#Count number of each value in GenHlth Column
#Would you say that in general your health is: scale 1-5 1 = excellent 2 = very good 3 = good 4 = fair 5 = poor
diabetes_df['GenHlth'].value_counts()

#percentage
diabetes_df.GenHlth.value_counts(normalize=True)

#Count number of each value in MentHlth Column
#Now thinking about your mental health, which includes stress, depression, and problems with 
#emotions, for how many days during the past 30 days was your mental health not good? scale 1-30 days
diabetes_df['MentHlth'].value_counts()


#percentage
diabetes_df.MentHlth.value_counts(normalize=True)

#### Buckets 
# 0 Days 
# 1-5 Days
# 6-10 Days
# 11-15 Days
# 16-20 Days
# 21-25 Days
# 26-30 Days

mhealth_bin_range = [-1, 5, 10, 15, 20, 25, 32]

mhealth_bins = [1, 2, 3, 4, 5, 6]

diabetes_df['Mental_Health_Range'] = pd.cut(diabetes_df['MentHlth'], mhealth_bin_range, labels=mhealth_bins)

diabetes_df.Mental_Health_Range.describe()

#reorder columns 
diabetes_df = diabetes_df[['Diabetes_Status', 'HighBP', 'HighChol', 'CholCheck', 'BMI_Range',
       'Smoker', 'Stroke', 'HeartDiseaseorAttack', 'PhysActivity', 'Fruits',
       'Veggies', 'HvyAlcoholConsump', 'AnyHealthcare', 'NoDocbcCost',
       'GenHlth', 'Mental_Health_Range', 'PhysHlth', 'DiffWalk', 'Sex', 'Age',
       'Education', 'Income', 'Diabetes_012', 'BMI', 'MentHlth']]

#post bucketing column validation 
diabetes_df.columns

#View data
diabetes_df.head()

#Table info 
diabetes_df.info()

#Count number of each value in PhysHlth Column
#Now thinking about your physical health, which includes physical illness and injury, for how many days 
#during the past 30 days was your physical health not good? scale 1-30 days
diabetes_df['PhysHlth'].value_counts()


#percentage
diabetes_df.PhysHlth.value_counts(normalize=True)

#bins
# 0 Days
# 1-5 Days
# 6-10 Days
# 11-15 Days
# 16-20 Days
# 21-25 Days
# 26-30 Days

phealth_bin_range = [-1, 5, 10, 15, 20, 25, 32]

phealth_bins = [1, 2, 3, 4, 5, 6]

diabetes_df['Physical_Health_Range'] = pd.cut(diabetes_df['PhysHlth'], phealth_bin_range, labels=phealth_bins)

diabetes_df.Physical_Health_Range.describe()

#post binning column validation 
diabetes_df.columns

#view data
diabetes_df.head()

#reorganize columns 
diabetes_df = diabetes_df[['Diabetes_Status', 'HighBP', 'HighChol', 'CholCheck', 'BMI_Range',
       'Smoker', 'Stroke', 'HeartDiseaseorAttack', 'PhysActivity', 'Fruits',
       'Veggies', 'HvyAlcoholConsump', 'AnyHealthcare', 'NoDocbcCost',
       'GenHlth', 'Mental_Health_Range', 'Physical_Health_Range', 'DiffWalk', 'Sex', 'Age',
       'Education', 'Income', 'Diabetes_012', 'BMI', 'MentHlth', 'PhysHlth']]

#table info
diabetes_df.info()

#Count number of each value in DiffWalk Column
#Do you have serious difficulty walking or climbing stairs? 0 = no 1 = yes
diabetes_df['DiffWalk'].value_counts()


#percentage
diabetes_df.DiffWalk.value_counts(normalize=True)

#Count number of each value in Sex Column
#0 = female 1 = male
diabetes_df['Sex'].value_counts()

#percentage
diabetes_df.Sex.value_counts(normalize=True)

#Count number of each value in Age Column
#13-level age category (_AGEG5YR see codebook) 1 = 18-24 9 = 60-64 13 = 80 or older
#link to codebook here: https://www.cdc.gov/brfss/annual_data/2015/pdf/codebook15_llcp.pdf
diabetes_df['Age'].value_counts()

#Percentage
diabetes_df.Age.value_counts(normalize=True)

#Count number of each value in Education Column
#Education level (EDUCA see codebook) scale 1-6 1 = Never attended school or only kindergarten 
#2 = Grades 1 through 8 (Elementary) 3 = Grades 9 through 11 (Some high school) 4 = Grade 12 or GED (High school graduate) 
#5 = College 1 year to 3 years (Some college or technical school) 6 = College 4 years or more (College graduate)
#link to codebook here: https://www.cdc.gov/brfss/annual_data/2015/pdf/codebook15_llcp.pdf
diabetes_df['Education'].value_counts()


#percentage
diabetes_df.Education.value_counts(normalize=True)

#Count number of each value in Income Column
#Income scale (INCOME2 see codebook) scale 1-8 1 = less than $10,000 5 = less than $35,000 8 = $75,000 or more
#link to codebook here https://www.cdc.gov/brfss/annual_data/2015/pdf/codebook15_llcp.pdf
diabetes_df['Income'].value_counts()


#percentage
diabetes_df.Income.value_counts(normalize=True)

#Preprocessing 

diabetes_df.columns

#drop original columns from binning 

#drop diabetes_012 
diabetes_df = diabetes_df.drop(["Diabetes_012"], axis=1)

print("done")


#drop BMI
diabetes_df = diabetes_df.drop(["BMI"], axis=1)

print("done")

#drop MentHlth
diabetes_df = diabetes_df.drop(["MentHlth"], axis=1)

print("done")

#drop PhysHlth
diabetes_df = diabetes_df.drop(["PhysHlth"], axis=1)

print("done")

diabetes_df.columns

#split target column out of data 
y = diabetes_df['Diabetes_Status']

y

x = diabetes_df.drop(['Diabetes_Status'], axis=1)

x

diabetes_df.to_csv('results/DataProcessingExtractFile-Raw Data.csv', encoding='utf-8')
print('done')
