==> archives/FInal_presentation_02032023/2_DiabetesPrediction_RandomForestModelGeneration.ipynb <==
### This file is run in the second step of the process.  It loads the data file created in step one and then generates the model.pkl file for predictions in the future. 

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import pickle
import requests
import json
from pathlib import Path
from sklearn import tree
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import GradientBoostingClassifier
from collections import Counter
from sklearn.datasets import make_blobs
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report

print("Pandas and Modeling dependencies loaded")

data = Path('./results/DataProcessingExtractFile-RawData.csv')
df = pd.read_csv(data)
df.head(10)

# Create Random Forest Model
X = df.copy()
X = df.drop(df.columns[[0]], axis=1)
X.head()

y = df["Diabetes_Status"].ravel()
y[:5]

# Splitting into Train and Test sets.
X_train, X_test, y_train, y_test = train_test_split(X, 
                                                    y,
                                                    random_state=78,
                                                    stratify=y)

# Create a random forest classifier.
rf_model = RandomForestClassifier(n_estimators=128, random_state=78) 

# Fitting the model
rf_model = rf_model.fit(X_train, y_train)

# Making predictions using the testing data.
predictions = rf_model.predict(X_test)
predictions

# Calculating the confusion matrix.
cm = confusion_matrix(y_test, predictions)

# Create a DataFrame from the confusion matrix.
cm_df = pd.DataFrame(
    cm, index=["Actual 0", "Actual 1"], columns=["Predicted 0", "Predicted 1"])

# Calculating the accuracy score.
acc_score = accuracy_score(y_test, predictions)
acc_score

# Displaying results
print("Confusion Matrix")
display(cm_df)
print(f"Accuracy Score : {acc_score}")
print("Classification Report")
print(classification_report(y_test, predictions))

# Calculate feature importance in the Random Forest model.
importances = rf_model.feature_importances_
importances

# We can sort the features by their importance.
sorted(zip(rf_model.feature_importances_, X.columns), reverse=True)

pickle.dump(rf_model, open('model.pkl','wb'))
