==> archives/FInal_presentation_02032023/3_Prediction_Machine_Learning_Code.ipynb <==
### This file is a python version of the code run in the JavaScript on Prediction site.  It can be used by reading in a json file that contains results similar to what is generated on the site to validate what the prediction model is doing. 

#import dependencies
import pandas as pd
import matplotlib.pyplot as plt
from pathlib import Path
from sklearn import tree
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import GradientBoostingClassifier
from collections import Counter
from sklearn.datasets import make_blobs
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report
from sklearn.linear_model import LogisticRegression
import pickle
import json
from pandas import json_normalize
import os
import csv
import numpy as np


loaded_model = pickle.load(open('./model.pkl', 'rb'))


#read in file from survey website
final_predict = Path('../site-dashboard/export/surveyExport.json')
prediction_df = pd.read_json(final_predict, lines=True)
prediction_df

#reset index 
prediction_df.reset_index(inplace=True)


prediction_df

#drop index
prediction_df.drop(columns=['index'])


#validate
predict_df = prediction_df[['HighBP', "HighChol", "CholCheck", "BMI_Range", "Smoker", "Stroke", "HeartDiseaseorAttack", "PhysActivity", "Fruits", "Veggies", "HvyAlcoholConsumP", "AnyHealthcare", "NoDocbcCost", "GenHlth", "Mental_Health_Range", "Physical_Health_Range", "DiffWalk", "Sex", "Age", "Education", "Income"]]
predict_df.head()

#set  predict dataframe to predict values  and run prediction against trained model 
y_pred = loaded_model.predict(predict_df)
#validate result completes 
y_pred

#set value to send back to website 
result = y_pred 

result

result_df=pd.DataFrame(result, columns=['result'])

result_df.to_csv('../results/results.csv')

result_df.to_json('../results/results.json')
